feat: cleanup: moved interfaces from layout component to interfaces folder, updated react-router-dom, removed unused portalpage
