+++ packages/app/package.json
@@ -24,7 +24,7 @@
     "@nl-portal/nl-portal-user-interface": "workspace:*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.21.3"
+    "react-router-dom": "^6.22.1"
   },
   "devDependencies": {
     "@types/jest": "^29.5.12",
