+++ packages/user-interface/package.json
@@ -96,6 +96,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-intl": "^6.6.2",
-    "react-router-dom": "^6.21.3"
+    "react-router-dom": "^6.22.1"
   }
 }
