+++ packages/user-interface/src/contexts/RouterOutletContext.tsx
@@ -1,4 +1,4 @@
-import { Paths } from "../components/Layout";
+import { Paths } from "../interfaces/paths";
 
 export interface RouterOutletContext {
   paths: Paths;
