+++ packages/user-interface/src/interfaces/navigation-item.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export interface NavigationItem {
+  titleTranslationKey: string;
+  path: string;
+  icon: ReactNode;
+  hasMessagesCount?: boolean;
+}
