+++ packages/user-interface/src/interfaces/paths.ts
@@ -1,0 +1,14 @@
+export interface Paths {
+  overview: string;
+  cases: string;
+  case: string;
+  tasks: string;
+  task: string;
+  notifications: string;
+  themes: string;
+  forms: string;
+  form: string;
+  account: string;
+  edit_account: string;
+  [key: string]: string | undefined;
+}
