--- packages/user-interface/src/interfaces/portal-page.ts
@@ -1,13 +1,0 @@
-import { ReactElement } from "react";
-
-export interface PortalPage {
-  path: string;
-  titleTranslationKey: string;
-  pageComponent: ReactElement;
-  isHome?: boolean;
-  showInMenu?: boolean;
-  showMessagesCount?: boolean;
-  icon?: ReactElement;
-  children?: Array<PortalPage>;
-  showLinkToParent?: boolean;
-}
